fix: use long for grandTotal to avoid implicit narrowing conversion in FindBlock, FindItem, FindEntity

+++ src/main/java/com/example/miscfeatures/command/FindBlock.java
@@ -204,7 +204,7 @@ public class FindBlock {
             }
         });
 
-        int grandTotal = 0;
+        long grandTotal = 0L;
         for (Map.Entry<SearchTarget, List<BlockPos>> entry : results.entrySet()) {
             SearchTarget target = entry.getKey();
             List<BlockPos> positions = entry.getValue();
